Fix page and per page assignment
In the class Pagination in constructor there was an assignment of variables page and per_page. They were assigned with the help of kwargs.get.
The problem was that kwargs.get was getting a per_page_param/page_parameter variable rather than the actual string specified in docstring ("page" and "per_page")

This commit fixes that issue.