update __init__.py to fix negative page issues
If you add an input tag in the html file with the attribute "name" the same as the name of page_parameter (default: name="page") it allows you to jump to the page written in the input. 
The problem is that if you input a negative number it will jump to the first page but pagination.info will be wrong and with negative numbers. 
Furthermore, even the indexes obtained with "{{ loop.index + pagination.skip }}" as per example in "example\templates\index.html" will be negative and wrong.

Defaulting self.page in Pagination.__init__() to 1 if it's negative fixes the issues.